some test

+++ execute/signal.c
@@ -6,18 +6,26 @@
 /*   By: sbendu <sbendu@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/19 11:26:10 by sbendu            #+#    #+#             */
-/*   Updated: 2022/06/20 23:20:01 by sbendu           ###   ########.fr       */
+/*   Updated: 2022/06/21 10:54:03 by sbendu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "execute.h"
 
+void	display_prompt(void)
+{	
+	write(1, "\n", 1);
+	rl_replace_line("", 0);
+	rl_on_new_line();
+	rl_redisplay();
+}
+
 void	sighandler(int signal)
 {
 	if (signal == SIGINT)
-	{
 		handle_ctrl_c(signal, NULL);
-	}
+	if (signal == SIGQUIT)
+		handle_ctrl_qu(signal, NULL);
 }
 
 void	handle_ctrl_c(int signal, int *ptr)
@@ -32,28 +40,41 @@ void	handle_ctrl_c(int signal, int *ptr)
 		if (saved == NULL)
 		{
 			printf("\n");
-			rl_on_new_line();
-			rl_replace_line("", 1);
-			rl_redisplay();
+			printDir();
+		}
+		else
+		{
+			i = -1;
+			while (saved[++i] != 0)
+			{
+				printf("pid is %d\n", saved[i]);
+				kill(saved[i], SIGTERM);
+			}
+			// printf("\n");
+		}
+	}
+}
+
+void	handle_ctrl_qu(int signal, int *ptr)
+{
+	static int	*saved = NULL;
+	int			i;
+
+	if (signal == 5000)
+		saved = ptr;
+	if (signal == SIGINT)
+	{
+		if (saved == NULL)
+		{
+			
 		}
 		else
 		{
 			i = -1;
 			while (saved[++i] != 0)
 				kill(saved[i], SIGTERM);
+			printf("Quit\n");
+			printDir();
 		}
 	}
 }
-
-// int	handle_ctrl_d(const int signal, int *ptr)
-// {
-// 	static int	*saved = NULL;
-// 	int			i;
-// 	if (signal == 5000)
-// 		saved = ptr;
-// 	if (signal == EOF)
-// 	{
-// 		if (saved == NULL)
-// 			;
-// 	}
-// }
